backend/requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
azure-common==1.1.28
azure-core==1.24.2
azure-storage-blob==12.14.1
python-multipart==0.0.6
aiohttp==3.8.5
python-dotenv==1.0.0
